requirements.txt
psutil==5.9.8
docker==7.0.0
psycopg2-binary
python-dotenv
requests
SQLAlchemy
openai
